refactored helper functions to selectors
